[update] redis 설정 파일 하나로 통합
